QOL Changes: - Split up main app into with database and form specific files. - Reworked app config so that db.init would come after app.config. - Added db.create_all to construct tables. - Added comments - Updated git ignore to exclude pycache files
