+++ .gitignore
@@ -1,2 +1,8 @@
+# venv
 py312_env\*
 py312_env/*
+
+# Byte-compiled / optimized / DLL files
+__pycache__/
+*.py[cod]
+*$py.class
