+++ forms.py
@@ -1,0 +1,23 @@
+from flask_wtf import FlaskForm
+from wtforms import StringField, PasswordField, SubmitField
+from wtforms.validators import InputRequired, Length, ValidationError
+
+# DB Imports
+from dbConfig import User
+
+
+class RegisterForm(FlaskForm):
+    username = StringField(validators=[InputRequired(), Length(min=4, max=20)], render_kw={"placeholder":"Username"}, )
+    password = PasswordField(validators=[InputRequired(), Length(min=8, max=20)], render_kw={"placeholder":"Password"})
+    submit = SubmitField("Register")
+
+    def validate_username(self, username):
+        exists = User.query.filter_by(username=username.data).first()
+        if exists:
+            raise ValidationError("Username already exists, please choose a different one.")
+        
+
+class LoginForm(FlaskForm):
+    username = StringField(validators=[InputRequired(), Length(min=4, max=20)], render_kw={"placeholder":"Username"})
+    password = PasswordField(validators=[InputRequired(), Length(min=8, max=20)], render_kw={"placeholder":"Password"})
+    submit = SubmitField("Login")
